Try to improve the intro
Just added a link from the website. It will be live relatively soon, and I know lots of people are going to be clicking it thinking it is something it is not. I often think a FAQ is going to be from the perspective of "I maybe want to use this, what are all the questions I should ask" which it sometimes is.

Anyway, point is, I tried to make the intro clearly state the intended audience. From there, I tried to help route people to places where they can get additional help.